% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FractalTools.R
\docType{package}
\name{FractalTools-package}
\alias{FractalTools}
\alias{FractalTools-package}
\title{FractalTools: Fractality-based tools}
\description{
Contains some tools based on the fractality concept such as
Local-fractality and cross local fractality.
}
\details{
The algorithm does not deal with missing values, please make 
sure to remove them. Since the algorithm is based on pairwise distances, 
and according to the computing power of your machine, large number of
data points can take much time and needs more memory.
}
\note{
This R code was developed and used for the following research: "Title"
 presented at the European Geosciences Union General Assembly (EGU 2018).
}
\references{
Mikhail Kanevski, Mario G. Pereira, Local fractality: The case of forest fires
 in Portugal, In Physica A,
 Volume 479, 2017, Pages 400-410, ISSN 0378-4371.
}
\author{
Mohamed Laib \email{Mohamed.Laib@unil.ch}  and

Mikhail Kanevski \email{Mikhail.Kanevski@unil.ch},
}
