% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalSB.R
\name{GlobalSB}
\alias{GlobalSB}
\title{Global fractality by fixing a radial (Sand-Box method)}
\usage{
GlobalSB(data, rad)
}
\arguments{
\item{data}{Data of class: \code{matrix} or \code{data.frame}.}

\item{rad}{Vector containning values of the radial.}
}
\value{
A list containing
 \itemize{
  \item \code{Fdim} Computed fractal dimension.
  \item \code{SB} Data frame of log (radial) and the log (number of points).
  }
}
\description{
Calculate the fractal dimension of a dataset by using the Sand-Box method
}
\examples{

\dontrun{
data <- unidat(n=10000, d=20)
data<-apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.7,0.9,0.01)
A<-GlobalSB(data, rad)
A$Fdim
plot(A[[2]][,1],A[[2]][,2],type="b",pch=16,lwd=2,xlab="log(Rad)",ylab="log(N)",
      main="Sandbox plot",  cex.main=1.5, las=1, axes=F)
axis(1)
axis(2)
grid()
Regline <- lm(A[[2]][,2]~A[[2]][,1])
abline(Regline, col="red", lwd=2)
legend("bottomright",paste("Fractal dimension:", round(A$Fdim,2)), bty="n")


#### IDmining examples #####
library(IDmining)

#### SwissRoll dataset ####

data<-SwissRoll(N=1000)
data<-apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.1,0.31,0.01)
A<-GlobalSB(data, rad)
A$Fdim
plot(A[[2]][,1],A[[2]][,2],type="b",pch=16,lwd=2,xlab="log(Rad)",ylab="log(N)",
      main="Sandbox plot",  cex.main=1.5, las=1, axes=F)
axis(1)
axis(2)
grid()
Regline <- lm(A[[2]][,2]~A[[2]][,1])
abline(Regline, col="red", lwd=2)
legend("bottomright",paste("Fractal dimension:", round(A$Fdim,2)), bty="n")

#### Butterfly dataset ####
data <- Butterfly(1000)[,-9]
data<-apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.1,0.31,0.01)
A<-GlobalSB(data, rad)
A$Fdim
plot(A[[2]][,1],A[[2]][,2],type="b",pch=16,lwd=2,xlab="log(Rad)",ylab="log(N)",
      main="Sandbox plot",  cex.main=1.5, las=1, axes=F)
axis(1)
axis(2)
grid()
Regline <- lm(A[[2]][,2]~A[[2]][,1])
abline(Regline, col="red", lwd=2)
legend("bottomright",paste("Fractal dimension:",
         round(Regline$coefficient[2],2)), bty="n")

}

}
