% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalSB.R
\name{LocalSB}
\alias{LocalSB}
\title{Local fractality (Sand-Box method) by fixing radial}
\usage{
LocalSB(data, Rad)
}
\arguments{
\item{data}{Data of class: \code{matrix} or \code{data.frame}.}

\item{Rad}{Vector containning values of the radial.}
}
\value{
A data.frame contains: data, columns at each radials. The last
     presents the Fractal dimension of each points.
}
\description{
Calculate the fractal dimension of a dataset by using the Sand-Box method
}
\examples{

\dontrun{

#### Uniform variables ####

data <- unidat(n=10000, d=20)
data<-apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.1,1,0.01)
A<-LocalSB(data, rad)
A$Results$Fdim


#### IDmining examples #####
library(IDmining)

#### SwissRoll dataset ####

data<-SwissRoll(N=1000)
data<-apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.1,1,0.01)
A<-LocalSB(data, rad)
A$Results$Fdim
#### Butterfly dataset ####
data <- Butterfly(1000)[,-9]
data <- apply(data,MARGIN=2,FUN = function(X) (X - min(X))/diff(range(X)))
rad<-seq(0.1,0.31,0.01)
A<-LocalSB(data, rad)
A$Results$Fdim

}

}
