% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidat.R
\name{unidat}
\alias{unidat}
\title{Simulated uniform matrix}
\usage{
unidat(n=1000, d=20)
}
\arguments{
\item{n}{Number of points.}

\item{d}{Number of variables.}
}
\value{
A matrix of n*d
}
\description{
Gives a matrix of uniform variables
}
\examples{

\dontrun{
raw_data<-unidat(n=1000, d=10)
dim(raw_data)
}

}
