% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossLocalSB.R
\name{CrossLocalSB}
\alias{CrossLocalSB}
\title{Cross Local fractality (Sand-Box method) by fixing radial}
\usage{
CrossLocalSB(data1, data2, rad)
}
\arguments{
\item{data1}{First dataset. Data of class: \code{matrix} or 
\code{data.frame}.}

\item{data2}{Second dataset. Data of class: \code{matrix} or 
\code{data.frame}.}

\item{Rad}{Vector containning values of the radial.}
}
\value{
A data.frame contains: data, columns at each radials. The last
     presents the Fractal dimension of each points.
}
\description{
Calculate the cross local fractal dimension between two
 datasets by using the Sand-Box method.
}
\examples{

\dontrun{

}

}
